Small fix to the save settings prompt
I've added a separate check during the prompt to detect an intentional instruction to not save, in order to avoid ambiguity when responding to the user's input. I've also added a try catch to the saving process, just in case the program doesn't have permission to write to the disk for some reason.